+++ client/src/commonMain/kotlin/org/apptank/horus/client/connectivity/Network.kt
@@ -1,9 +1,20 @@
 package org.apptank.horus.client.connectivity
 
+import org.apptank.horus.client.config.HorusPreferences
+
 data class Network(
     val connections: List<NetworkConnection>
 ) {
     fun hasValidConnection(): Boolean {
+
+        if (HorusPreferences.offlineMode) {
+            return false
+        }
+
+        if (HorusPreferences.ignoreNetworkStatus) {
+            return true
+        }
+
         return connections.any { it.level.level >= ConnectionLevel.MEDIUM.level }
     }
 
@@ -12,6 +23,15 @@ data class Network(
     }
 
     fun isOfflineMode(): Boolean {
+
+        if (HorusPreferences.offlineMode) {
+            return true
+        }
+
+        if (HorusPreferences.ignoreNetworkStatus) {
+            return false
+        }
+
         return hasValidConnection().not()
     }